refactored dom classes of outliers
