+++ src/main/java/gr/uoi/cs/pythia/model/outlier/OutlierType.java
@@ -1,0 +1,5 @@
+package gr.uoi.cs.pythia.model.outlier;
+
+public enum OutlierType {
+	Z_SCORE
+}
