+++ src/main/java/gr/uoi/cs/pythia/patterns/outlier/IOutlierAlgo.java
@@ -6,6 +6,7 @@ import org.apache.spark.sql.Dataset;
 import org.apache.spark.sql.Row;
 
 import gr.uoi.cs.pythia.model.DatasetProfile;
+import gr.uoi.cs.pythia.model.outlier.OutlierResult;
 
 public interface IOutlierAlgo {
