--- src/main/java/gr/uoi/cs/pythia/patterns/outlier/OutlierType.java
@@ -1,5 +1,0 @@
-package gr.uoi.cs.pythia.patterns.outlier;
-
-public enum OutlierType {
-	Z_SCORE
-}
